requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
cffi==1.10.0
cryptography==1.9
idna==2.5
packaging==16.8
paramiko==2.1.2
pyasn1==0.2.3
pycparser==2.17
pyparsing==2.2.0
pysftp==0.2.9
PyYAML==3.12
six==1.10.0
